pyproject.toml
[project]
name = "tokenization-strategy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ai-sdk-python>=0.1.1",
    "datasets<4.0.0",
    "fugashi[unidic]>=1.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
